Clean redundant test
